+++ EpicLoot/EpicLoot.csproj
@@ -232,6 +232,7 @@
     <Compile Include="MagicItemEffects\ModifyHealthRegen.cs" />
     <Compile Include="MagicItemEffects\ModifyJumpStaminaUsage.cs" />
     <Compile Include="MagicItemEffects\ModifyParry.cs" />
+	<Compile Include="MagicItemEffects\ModifyParryForce.cs" />
     <Compile Include="MagicItemEffects\ModifyParryWindow.cs" />
     <Compile Include="MagicItemEffects\ModifyResistance.cs" />
     <Compile Include="MagicItemEffects\ModifyRunStaminaDrain.cs" />
