+++ EpicLoot/MagicItemEffects/ModifyParry.cs
@@ -3,29 +3,6 @@ using HarmonyLib;
 
 namespace EpicLoot.MagicItemEffects
 {
-	//public float GetDeflectionForce(int quality)
-	[HarmonyPatch(typeof(ItemDrop.ItemData), nameof(ItemDrop.ItemData.GetDeflectionForce), typeof(int))]
-	public static class ModifyParry_ItemData_GetDeflectionForce_Patch
-	{
-		public static void Postfix(ItemDrop.ItemData __instance, ref float __result)
-        {
-            var player = PlayerExtensions.GetPlayerWithEquippedItem(__instance);
-            var totalParryMod = 0f;
-			ModifyWithLowHealth.Apply(player, MagicEffectType.ModifyParry, effect =>
-            {
-                totalParryMod += MagicEffectsHelper.GetTotalActiveMagicEffectValueForWeapon(player, __instance, effect, 0.01f);
-            });
-
-			__result *= 1.0f + totalParryMod;
-            if (player != null && player.m_leftItem == null && MagicEffectsHelper.HasActiveMagicEffect(player, __instance, MagicEffectType.Duelist))
-			{
-				__result += __instance.GetDamage().GetTotalDamage() / 2 * MagicEffectsHelper.GetTotalActiveMagicEffectValueForWeapon(player, __instance, MagicEffectType.Duelist, 0.01f);
-			}
-
-			__result = (float) Math.Round(__result, 1);
-		}
-	}
-
 	//public override bool BlockAttack(HitData hit, Character attacker)
 	[HarmonyPatch(typeof(Humanoid), nameof(Humanoid.BlockAttack))]
 	public static class ModifyParry_Humanoid_BlockAttack_Patch
