+++ EpicLoot/translations.json
@@ -586,10 +586,15 @@
   "mod_epicloot_me_modifyblockpower_suffix1": "Stopping",
   "mod_epicloot_me_modifyblockpower_suffix2": "the Defender",
   "mod_epicloot_me_modifyparry_display": "Parry +{0:0.#}%",
-  "mod_epicloot_me_modifyparry_desc": "Increase the parry block armor, parry force, and parry bonus of the magic by +<b><color=yellow>X</color></b>%.",
+  "mod_epicloot_me_modifyparry_desc": "Increase the parry block armor and parry bonus of the magic item by +<b><color=yellow>X</color></b>%.",
   "mod_epicloot_me_modifyparry_prefix1": "Elusive",
   "mod_epicloot_me_modifyparry_prefix2": "Rebuking",
   "mod_epicloot_me_modifyparry_suffix1": "Repelling",
+  "mod_epicloot_me_modifyparryforce_display": "Parry Force +{0:0.#}%",
+  "mod_epicloot_me_modifyparryforce_desc": "Increase the parry force of the magic item by +<b><color=yellow>X</color></b>%.",
+  "mod_epicloot_me_modifyparryforce_prefix1": "Weighty",
+  "mod_epicloot_me_modifyparryforce_prefix2": "Ponderous",
+  "mod_epicloot_me_modifyparryforce_suffix1": "Tossing",
   "mod_epicloot_me_modifyarmor_display": "Armor +{0:0.#}%",
   "mod_epicloot_me_modifyarmor_desc": "Increase the armor value of the magic equipment by +<b><color=yellow>X</color></b>%.",
   "mod_epicloot_me_modifyarmor_prefix1": "Heavy",
